Adapter Simplex

+++ src/main/java/ar/edu/utn/frba/dds/dispositivo/DispositivoInteligenteAAcondicionado.java
@@ -5,7 +5,7 @@ import ar.edu.utn.frba.dds.actuador.ActuadorAAcondicionado;
 public class DispositivoInteligenteAAcondicionado extends DispositivoInteligente {
     private int temperatura;
 
-    public DispositivoInteligenteAAcondicionado(String nombre, int consumoHora, String estado) {
+    public DispositivoInteligenteAAcondicionado(String nombre, double consumoHora, String estado) {
         super(nombre, consumoHora, estado,90,360);
         ActuadorAAcondicionado actuador = new ActuadorAAcondicionado(this);
         this.setActuador(actuador);
