+++ src/main/java/ar/edu/utn/frba/dds/dispositivo/DispositivoInteligenteHeladera.java
@@ -6,7 +6,7 @@ public class DispositivoInteligenteHeladera extends DispositivoInteligente {
 
     private int intensidad;
 
-    public DispositivoInteligenteHeladera(String nombre, int consumoHora, String estado) {
+    public DispositivoInteligenteHeladera(String nombre, double consumoHora, String estado) {
         super(nombre, consumoHora, estado,90,360);
         ActuadorHeladera actuador = new ActuadorHeladera(this);
         this.setActuador(actuador);
