+++ src/main/java/ar/edu/utn/frba/dds/dispositivo/DispositivoInteligenteLavarropas.java
@@ -5,7 +5,7 @@ import ar.edu.utn.frba.dds.actuador.ActuadorLavarropas;
 public class DispositivoInteligenteLavarropas extends DispositivoInteligente {
 
 
-    public DispositivoInteligenteLavarropas(String nombre, int consumoHora, String estado) {
+    public DispositivoInteligenteLavarropas(String nombre, double consumoHora, String estado) {
         super(nombre, consumoHora, estado,6,30);
         ActuadorLavarropas actuador = new ActuadorLavarropas(this);
         this.setActuador(actuador);
