+++ src/main/java/ar/edu/utn/frba/dds/dispositivo/DispositivoInteligentePC.java
@@ -5,7 +5,7 @@ import ar.edu.utn.frba.dds.actuador.ActuadorPC;
 public class DispositivoInteligentePC extends DispositivoInteligente {
 
 
-    public DispositivoInteligentePC(String nombre, int consumoHora, String estado) {
+    public DispositivoInteligentePC(String nombre, double consumoHora, String estado) {
         super(nombre, consumoHora, estado,60,360);
         ActuadorPC actuador = new ActuadorPC(this);
         this.setActuador(actuador);
